Share described enum base
